requirements/base.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.1
django-geoposition==0.2.2
django-toolbelt==0.0.1
envvars==0.3.0
geopy==1.10.0
google-api-python-client==1.4.0
googlemaps==2.2
gunicorn==19.3.0
httplib2==0.9.1
oauth2client==1.4.10
psycopg2==2.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
requests==2.7.0
rsa==3.1.4
simplejson==3.7.1
six==1.9.0
static3==0.6.0
uritemplate==0.6
whitenoise==1.0.6